zero: core, fixed major heap memory leaks & logic in producer consumer recovery path code zero: scheduler refit with core tech instead & upgrades zero: added connection reset detection & propogation on net sockets (investigate connection reset from socket?) zero: Q reworked atomic set (superfast & test ok) using a CAS bloom filter instead (memory hungry) zero: Q now supports value types at the cost of memory

+++ zero.test/core/patterns/queue/IoQTest.cs
@@ -196,7 +196,7 @@ namespace zero.test.core.patterns.queue
                             success++;
                         }
                         
-                        maxIdx = Math.Max(maxIdx, bag.Count);
+                        maxIdx = (int)Math.Max(maxIdx, bag.Count);
                     }
 
                     Interlocked.Increment(ref @this.SpamTestAsyncThreadsDone);
